requirements.txt
# Core Deep Learning (Latest versions compatible with Python 3.13)
torch>=2.6.0
torchvision>=0.21.0

# Transformers and NLP
transformers>=4.40.0
tokenizers>=0.19.0

# CLIP
git+https://github.com/openai/CLIP.git

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pillow>=10.0.0
datasets>=2.14.0

# NLP Utilities
nltk>=3.8.1
textstat>=0.7.3
vaderSentiment>=3.3.2

# Machine Learning
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Progress Bars
tqdm>=4.66.0

# Jupyter (optional, for notebooks)
jupyter>=1.0.0
ipywidgets>=8.1.0